Refactor: Simplify runtime configuration interface
Remove complex runtime objects and dead code, providing a simplified interface with essential helpers: env, dns, and url.
