+++ packages/core/src/runtime-config.ts
@@ -1,100 +1,61 @@
-import type { TsOpsConfig, EnvValue, SecretRef, ConfigMapRef, ExtractNamespaceVarsFromConfig } from './types.js'
+import type { TsOpsConfig, DNSType, ExtractNamespaceVarsFromConfig } from './types.js'
 import { createConfigResolver } from './config/resolver.js'
-import { isSecretRef, isConfigMapRef } from './types.js'
 
 /**
- * Resolved environment variables for a single app.
- * All functions are evaluated, secrets/configMaps are resolved to actual values.
+ * Runtime helper context for building environment variables and URLs
  */
-export interface ResolvedAppEnv {
-  /** Resolved environment variables as key-value pairs */
-  env: Record<string, string>
-  /** Internal Kubernetes DNS endpoint */
-  internalEndpoint: string
-  /** Kubernetes service name */
-  serviceName: string
-  /** Resolved external host (if configured via network) */
-  host?: string
-  /** Docker image reference */
-  image: string
-  /** Resolved secrets (if any) */
-  secrets: Record<string, Record<string, string>>
-  /** Resolved configMaps (if any) */
-  configMaps: Record<string, Record<string, string>>
-}
-
-/**
- * Runtime configuration for all apps in a specific namespace.
- * All dynamic values are resolved and ready to use.
- */
-export type RuntimeConfig<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>> = {
-  /** Current namespace being configured */
+export interface RuntimeHelperContext<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>> {
+  /** Current namespace */
   namespace: Extract<keyof TConfig['namespaces'], string>
   /** Project name */
   project: TConfig['project']
   /** Namespace variables */
   namespaceVars: ExtractNamespaceVarsFromConfig<TConfig>
-  /** Resolved app configurations */
-  apps: {
-    [K in Extract<keyof TConfig['apps'], string>]: ResolvedAppEnv
-  }
+  /** External hosts mapping */
+  externalHosts: Record<string, string>
+  /** Apps configuration */
+  appsConfig: Record<string, any>
 }
 
 /**
- * Creates a runtime configuration for a specific namespace.
- * All functions in the config are evaluated and resolved to actual values.
- * 
- * @param config - TsOps configuration object
- * @param namespace - Target namespace to resolve config for
- * @returns Runtime configuration with all values resolved
- * 
- * @example
- * ```typescript
- * import { createRuntimeConfig } from '@tsops/core'
- * import config from './tsops.config'
- * 
- * // Create runtime config for 'dev' namespace
- * const runtime = createRuntimeConfig(config, 'dev')
- * 
- * // Access resolved environment variables
- * const apiUrl = runtime.apps['worken-api'].env.DATABASE_URL
- * 
- * // Get internal Kubernetes endpoint
- * const endpoint = runtime.apps['worken-api'].internalEndpoint
- * // => 'http://worken-api:3000'
- * 
- * // Type-safe access to all apps
- * for (const [appName, appConfig] of Object.entries(runtime.apps)) {
- *   console.log(`${appName}: ${appConfig.internalEndpoint}`)
- * }
- * ```
+ * Simplified runtime configuration with only essential methods
  */
-export function createRuntimeConfig<
-  TConfig extends TsOpsConfig<any, any, any, any, any, any, any>
->(
+export interface RuntimeConfig<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>> {
+  /** Current namespace */
+  namespace: Extract<keyof TConfig['namespaces'], string>
+  /** Project name */
+  project: TConfig['project']
+  /** Namespace variables */
+  namespaceVars: ExtractNamespaceVarsFromConfig<TConfig>
+  /** Get environment variables for an app */
+  getEnv(appName: Extract<keyof TConfig['apps'], string>): Record<string, string>
+  /** Generate DNS name */
+  dns(appName: Extract<keyof TConfig['apps'], string>, type: DNSType): string
+  /** Generate complete URL */
+  url(appName: Extract<keyof TConfig['apps'], string>, type: DNSType, options?: { protocol?: 'http' | 'https' }): string
+}
+
+/**
+ * Creates runtime helper functions for a specific namespace
+ */
+function createRuntimeHelpers<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>>(
   config: TConfig,
   namespace: Extract<keyof TConfig['namespaces'], string>
-): RuntimeConfig<TConfig> {
+) {
   const resolver = createConfigResolver(config)
-  
-  // Get namespace variables
   const namespaceVars = config.namespaces[namespace] as ExtractNamespaceVarsFromConfig<TConfig>
   
-  // Resolve all apps for this namespace
-  const apps = {} as RuntimeConfig<TConfig>['apps']
-  const appEntries = resolver.apps.select()
-  
-  // First pass: collect all external hosts and app configs
+  // Collect all external hosts and app configs
   const externalHosts: Record<string, string> = {}
   const appsConfig: Record<string, any> = {}
   
+  const appEntries = resolver.apps.select()
   for (const [appName, app] of appEntries) {
-    // Skip apps that shouldn't be deployed to this namespace
     if (!resolver.apps.shouldDeploy(app, namespace as string)) {
       continue
     }
     
-    // Store app config for url helper
+    // Store app config
     appsConfig[appName] = app
     
     // Create temporary context to resolve network
@@ -114,129 +75,144 @@ export function createRuntimeConfig<
     }
   }
   
-  // Second pass: resolve all apps with full context
-  for (const [appName, app] of appEntries) {
-    // Skip apps that shouldn't be deployed to this namespace
-    if (!resolver.apps.shouldDeploy(app, namespace as string)) {
-      continue
-    }
-    
-    // Create context for resolution with external hosts and apps config
-    const context = resolver.namespaces.createHostContext(namespace as string, { 
-      appName, 
-      externalHosts, 
-      appsConfig 
-    })
-    
-    // Resolve all app properties
-    const host = externalHosts[appName]
-    const envRaw = resolver.apps.resolveEnv(app, namespace as string, context)
-    const secrets = resolver.apps.resolveSecrets(app, namespace as string, context)
-    const configMaps = resolver.apps.resolveConfigMaps(app, namespace as string, context)
-    const image = app.image || resolver.images.buildRef(appName)
-    
-    // Convert env values to plain strings
-    const env = resolveEnvToStrings(envRaw, secrets, configMaps)
-    
-    // Get service name and internal endpoint
-    const serviceName = resolver.project.serviceName(appName)
-    const internalEndpoint = buildInternalEndpoint(
-      serviceName,
-      namespace as string,
-      app.ports?.[0]?.targetPort || app.ports?.[0]?.port || 3000
-    )
-    
-    apps[appName as Extract<keyof TConfig['apps'], string>] = {
-      env,
-      internalEndpoint,
-      serviceName,
-      host,
-      image,
-      secrets,
-      configMaps
-    }
-  }
-  
-  return {
+  const context: RuntimeHelperContext<TConfig> = {
     namespace,
     project: config.project,
     namespaceVars,
-    apps
+    externalHosts,
+    appsConfig
   }
-}
-
-/**
- * Helper to resolve env values to plain strings.
- * Handles SecretRef and ConfigMapRef by looking up actual values.
- */
-function resolveEnvToStrings(
-  envRaw: Record<string, EnvValue> | SecretRef | ConfigMapRef,
-  secrets: Record<string, Record<string, string>>,
-  configMaps: Record<string, Record<string, string>>
-): Record<string, string> {
-  const result: Record<string, string> = {}
   
-  // Handle envFrom (entire secret/configMap)
-  if (isSecretRef(envRaw)) {
-    const secretData = secrets[envRaw.secretName]
-    if (secretData) {
-      Object.assign(result, secretData)
+  /**
+   * Generate DNS name for different types of resources
+   */
+  const dns = (app: Extract<keyof TConfig['apps'], string>, type: DNSType): string => {
+    switch (type) {
+      case 'service':
+        return app
+      case 'ingress':
+        return context.externalHosts[app] || app
+      case 'cluster':
+      default:
+        return `${app}.${context.namespace}.svc.cluster.local`
     }
+  }
+  
+  /**
+   * Generate complete URL for different types of resources with automatic port resolution
+   */
+  const url = (app: Extract<keyof TConfig['apps'], string>, type: DNSType, options?: { protocol?: 'http' | 'https' }): string => {
+    const { protocol = 'http' } = options || {}
+    
+    // Get the first port from the app's configuration
+    const appConfig = context.appsConfig[app]
+    const firstPort = appConfig?.ports?.[0]?.port || 80
+    
+    // Get the DNS name
+    const hostname = dns(app, type)
+    
+    // Build the complete URL
+    return `${protocol}://${hostname}:${firstPort}`
+  }
+  
+  /**
+   * Build environment variables for an app using helper functions
+   */
+  const env = (appName: Extract<keyof TConfig['apps'], string>, envBuilder: (helpers: {
+    dns: typeof dns
+    url: typeof url
+    namespace: typeof context.namespace
+    project: typeof context.project
+    cluster: { name: string; apiServer: string; context: string }
+    appName: string
+    secret: (secretName: string, key?: string) => any
+    configMap: (configMapName: string, key?: string) => any
+  } & ExtractNamespaceVarsFromConfig<TConfig>) => Record<string, any>): Record<string, string> => {
+    // Create secret and configMap helpers (simplified - return placeholder values)
+    const secret = (secretName: string, key?: string) => {
+      if (key) {
+        return `secret:${secretName}:${key}`
+      }
+      return `secret:${secretName}`
+    }
+    
+    const configMap = (configMapName: string, key?: string) => {
+      if (key) {
+        return `configmap:${configMapName}:${key}`
+      }
+      return `configmap:${configMapName}`
+    }
+    
+    // Create full context for env builder
+    const fullContext = {
+      dns,
+      url,
+      namespace: context.namespace,
+      project: context.project,
+      cluster: { name: '', apiServer: '', context: '' }, // TODO: get from cluster config
+      appName,
+      secret,
+      configMap,
+      ...context.namespaceVars
+    }
+    
+    const envRaw = envBuilder(fullContext)
+    
+    // Convert to plain strings
+    const result: Record<string, string> = {}
+    for (const [key, value] of Object.entries(envRaw)) {
+      if (typeof value === 'string') {
+        result[key] = value
+      } else if (typeof value === 'number' || typeof value === 'boolean') {
+        result[key] = String(value)
+      }
+    }
+    
     return result
   }
   
-  if (isConfigMapRef(envRaw)) {
-    const configMapData = configMaps[envRaw.configMapName]
-    if (configMapData) {
-      Object.assign(result, configMapData)
-    }
-    return result
+  return {
+    dns,
+    url,
+    env,
+    context
   }
-  
-  // Handle individual env vars
-  for (const [key, value] of Object.entries(envRaw)) {
-    if (typeof value === 'string') {
-      result[key] = value
-    } else if (isSecretRef(value)) {
-      // Resolve secret key reference
-      const secretData = secrets[value.secretName]
-      if (secretData && value.key) {
-        result[key] = secretData[value.key]
-      }
-    } else if (isConfigMapRef(value)) {
-      // Resolve configMap key reference
-      const configMapData = configMaps[value.configMapName]
-      if (configMapData && value.key) {
-        result[key] = configMapData[value.key]
-      }
-    }
-  }
-  
-  return result
 }
 
 /**
- * Builds internal Kubernetes service endpoint.
- * Uses service name directly for cluster-internal communication.
+ * Creates a simplified runtime configuration
  */
-function buildInternalEndpoint(
-  serviceName: string,
-  namespace: string,
-  port: number | string
-): string {
-  const portNum = typeof port === 'string' ? port : port
-  return `http://${serviceName}:${portNum}`
+export function createRuntimeConfig<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>>(
+  config: TConfig,
+  namespace: Extract<keyof TConfig['namespaces'], string>
+): RuntimeConfig<TConfig> {
+  const helpers = createRuntimeHelpers(config, namespace)
+  
+  return {
+    namespace: helpers.context.namespace,
+    project: helpers.context.project,
+    namespaceVars: helpers.context.namespaceVars,
+    getEnv: (appName) => {
+      const app = helpers.context.appsConfig[appName]
+      if (!app || !app.env) {
+        return {}
+      }
+      
+      // If env is a function, call it with helpers
+      if (typeof app.env === 'function') {
+        return helpers.env(appName, app.env)
+      }
+      
+      // If env is an object, return as is
+      return app.env
+    },
+    dns: helpers.dns,
+    url: helpers.url
+  }
 }
 
 /**
- * Type helper to extract runtime config type from a TsOps config.
- * Useful for type annotations.
- * 
- * @example
- * ```typescript
- * import config from './tsops.config'
- * type MyRuntimeConfig = InferRuntimeConfig<typeof config>
- * ```
+ * Type helper for inferring runtime config type
  */
 export type InferRuntimeConfig<TConfig extends TsOpsConfig<any, any, any, any, any, any, any>> = 
   RuntimeConfig<TConfig>
